Profile view created and better navigation system implemented
